add tests for strategy
